all_nr_spsp_cvar_params

+++ portfolio_programming/simulation/wp_eg.pyx
@@ -443,7 +443,8 @@ class NIRExpPortfolio(WeightPortfolio, NIRUtility):
         S = self.column_stochastic_matrix(self.n_symbol,
                                            virtual_expert_weights.values)
         eigs, eigvs = np.linalg.eig(S)
-        normalized_new_weights = eigvs[:, 0] / eigvs[:, 0].sum()
+        normalized_new_weights = (eigvs[:, 0] / eigvs[:, 0].sum()).astype(
+            np.float64)
 
         # record modified strategies of today
         self.virtual_expert_decision_xarr.loc[
